# docs/swagger.yaml
basePath: /
definitions:
  handler.CreateNoteRequest:
    properties:
      content:
        type: string
      title:
        maxLength: 255
        type: string
    required:
    - content
    - title
    type: object
  handler.ListNotesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.NoteResponse'
        type: array
      meta:
        $ref: '#/definitions/model.PaginationMeta'
    type: object
  handler.UpdateNoteRequest:
    properties:
      content:
        type: string
      title:
        maxLength: 255
        type: string
    type: object
...
  model.ErrorDetail:
    properties:
      code:
        type: string
      details:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
  model.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/model.ErrorDetail'
    type: object
...
  model.NoteResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  model.PaginationMeta:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Production-ready Go web service starter kit with JWT auth, Argon2id,
    SQLC, and Chi.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: StarterKit API
  version: "1.0"
paths:
...
  /api/v1/notes:
    get:
      description: List notes for the authenticated user (or all notes for admin)
      parameters:
      - default: 20
        description: Number of items to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListNotesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List notes
      tags:
      - notes
    post:
      consumes:
      - application/json
      description: Create a new note for the authenticated user
      parameters:
      - description: Note details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateNoteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.NoteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new note
      tags:
      - notes
  /api/v1/notes/{id}:
    delete:
      description: Soft-delete a note
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a note
      tags:
      - notes
    get:
      description: Get a single note by ID
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.NoteResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a note
      tags:
      - notes
    put:
      consumes:
      - application/json
      description: Update an existing note
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      - description: Note updates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.NoteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a note
      tags:
      - notes
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"